\name{accepted_values}

\alias{accepted_values}

\title{
Get accepted terms
}

\description{
Get a data.frame with accepted terms (variable names)
}

\usage{
accepted_values(type, group)
}
 
\arguments{
  \item{type}{character. Either "records" (for the actual data) or "metadata" (for the metadata)}
  \item{group}{character. The group name}
}

\value{
character
}

