\name{data_from_uri}

\alias{data_from_uri}


\title{
Download data associated with a URI
}

\description{
Download data associated with a URI
}


\usage{
data_from_uri(uri, path, overwrite=FALSE)
}
 
\arguments{
  \item{uri}{character, a URI}
  \item{path}{character, the path}
  \item{overwrite}{logical. Should existing data be overwritten?}
}

\value{
character (filenames)
}


