\name{draft}

\alias{draft}

\title{
Generate a draft Carob script
}

\description{
Generate a draft Carob script to get you started
}


\usage{
draft(uri, path, group="draft", overwrite=FALSE)
}
 
\arguments{
  \item{uri}{character}
  \item{path}{character. Path to the carob repo}
  \item{group}{character.}
  \item{overwrite}{logical}
}

\value{
character, the filename of the script
}

