\name{geocode_duplicates}

\alias{geocode_duplicates}


\title{
Find coordinates for duplicated geographic locations
}

\description{
Find coordinates for duplicated geographic locations
}


\usage{
geocode_duplicates(x, vars=NULL)
}
 
\arguments{
  \item{x}{data.frame}
  \item{vars}{variable names to use in identifying duplicates, if not supplied all the available information is used}
}

\value{
data.frame
}


