\name{get_description}

\alias{get_license}
\alias{get_description}
\alias{get_authors}
\alias{get_title}

\title{
Extract specific metadata items from a list with metadata as returned by \code{\link{get_metadata}}
}

\description{
Extract specific metadata items
}


\usage{
get_license(x)

get_description(x)

get_authors(x)

get_title(x)
}
 
\arguments{
  \item{x}{list with metadata as returned by \code{\link{get_metadata}}}
}

\value{
character
}
