\name{get_packages}

\alias{get_packages}

\title{
Install required R packages
}

\description{
The function finds the R packages used in the scripts, so it may not be 100% accurate. 
}


\usage{
get_packages(path, group="")
}
 
\arguments{
  \item{path}{character. The path where the scripts are stored}
  \item{group}{character. optional, a group name}
}


