\name{get_terms}

\alias{get_terms}

\title{
Get accepted terms
}

\description{
Get a data.frame with accepted terms (variable names)
}


\usage{
get_terms(type, group, path)
}
 
\arguments{
  \item{type}{character. Either "records" (for the actual data) or "dataset" (for the metadata)}
  \item{group}{character. The group name}
  \item{path}{character. The path to the Carob repo}
}

\value{
character
}

