\name{is_excel}

\alias{is_excel}

\title{
Is this an excel file?
}

\description{
Check if a file is an excel file. This is useful as there are files with extension .csv that are actually excel files, and there are files with extension .xls(x) that are actually .csv files
}


\usage{
is_excel(f)
}
 
\arguments{
  \item{f}{character. filename}
}

\value{
logical
}
