\name{get_packages}

\alias{get_packages}

\title{
Install required packages
}

\description{
Install the R packages required to process the Carob datasets. This requires the \code{revtools} package.
}


\usage{
get_packages(path)
}
 
\arguments{
  \item{path}{the path to the Carob repo}
}

\value{
None
}

