\name{read.dta}

\alias{read.dta}

\title{
Read an dta file
}

\description{
Read an dta file and return a data.frame
}


\usage{
read.dta(f, unlabel=TRUE, ...)
}
 
\arguments{
  \item{f}{character (filename)}
  \item{unlabel}{logical}
  \item{...}{additional arguments passed on to \code{\link[haven]{read_dta}}}.
}

\value{
data.frame
}

