\name{get_references}

\alias{get_references}

\title{
Get dataset references
}

\description{
Get the references associated with each dataset in a compilation.
}


\usage{
get_references(x, path, format=TRUE)
}
 
\arguments{
  \item{x}{data.frame}
  \item{path}{the path to the Carob repo}
  \item{format}{logical. If \code{TRUE} the reference is returned in a formatted form that can be used in a list of references in an article}
}

\value{
character
}
