\name{update_terms}

\alias{update_terms}

\title{
Update the accepted terms
}

\description{
Update the accepted terms (variables and values) from the "terminag" repo. You must have permission to write to your "carobiner" package installation. If this does not work, you can update the package itself, as this should always have the latest version of the terms. 
}


\usage{
update_terms()
}
 
\value{
none
}

