\name{usr_pwd}

\alias{usr_pwd}

\title{
Get a username and password
}

\description{
Get a username and password
}


\usage{
usr_pwd(path, protocol)
}
 
\arguments{
  \item{path}{character. The path to the Carob repo}
  \item{protocol}{character. protocol name such as "LSMS"}
}

\value{
\code{data.frame} or \code{NULL}
}

