\name{get_accepted_values}

\alias{get_accepted_values}

\title{
Get accepted values for a variable (term)
}

\description{
Get a data.frame with accepted values for a (variable name). 
}


\usage{
get_accepted_values(term, path)
}
 
\arguments{
  \item{term}{character. e.g. "country", "crop", or "fertilizer_type"}
  \item{path}{character. The path to the Carob repo}
}

\value{
character
}

