\name{update_terms}

\alias{update_terms}

\title{
Update the accepted terms
}

\description{
Update the accepted terms (variables and values) from the "terminag" repo. You must have permission to write to your "carobiner" package installation. 
}


\usage{
update_terms(quiet=FALSE, force = FALSE, local_terms = NULL)
}
 
\arguments{
  \item{quiet}{logical. If \code{TRUE} a message is printed}
  \item{force}{logical}
  \item{local_terms}{character (path)}
} 
 
\value{
none
}

